fix: address sub-threshold code review findings
- Split registerHooks into mergeHooks + registerHooks (both under 50 lines)
- Deduplicate hooks by basename (not full path) to prevent accumulation on upgrades
- Fix TOCTOU race: use umask for atomic file creation with restrictive permissions
- Extract getUsage() to cli-usage.js (cli.js now 289 lines, under 300 limit)
- Handle --key=value syntax in CLI arg parser
- Sanitize SESSION_ID to prevent path traversal in event file paths
- Fix jq slice syntax ([:500] -> [0:500]) for broader compatibility
- Cap event JSONL files at 5MB to prevent unbounded growth
- Add Phase 2 forward-reference comment for stop-hook.sh in preuninstall
